actually fix build error

+++ db/assistants.go
@@ -75,7 +75,7 @@ func NewAccountWithID(uid string, resetcount int64) netobj.Player {
 	eventState := netobj.DefaultEventState()
 	battleState := netobj.DefaultBattleState()
 	disallowInactivePurge := false
-	lastLoginPlatformID := 0
+	lastLoginPlatformID := int64(0)
 	return netobj.NewPlayer(
 		uid,
 		username,
